+++ package.json
@@ -13,7 +13,7 @@
     "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.1",
-    "styled-component": "^2.8.0",
+    "styled-components": "^6.1.1",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
