+++ src/index.js
@@ -7,7 +7,9 @@ import { store } from "./store/index";
 import Modal from "react-modal";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
+
 Modal.setAppElement("#root");
+
 root.render(
   <React.StrictMode>
     <Provider store={store}>
